[#36] Refactor: 코드 정리
