+++ src/main/java/umc/GrowIT/Server/web/controller/AuthController.java
@@ -1,5 +1,6 @@
 package umc.GrowIT.Server.web.controller;
 
+import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
@@ -10,9 +11,9 @@ import umc.GrowIT.Server.web.dto.UserDTO.UserRequestDTO;
 import umc.GrowIT.Server.web.dto.UserDTO.UserResponseDTO;
 import umc.GrowIT.Server.service.userService.UserCommandService;
 
+@Tag(name = "Auth", description = "인증 관련 API")
 @RestController
 @RequiredArgsConstructor
-
 public class AuthController implements AuthSpecification {
 
     private final UserCommandService userCommandService;
