+++ src/main/java/umc/GrowIT/Server/web/controller/ChallengeController.java
@@ -9,26 +9,19 @@ import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
 import umc.GrowIT.Server.apiPayload.ApiResponse;
-import umc.GrowIT.Server.converter.ChallengeConverter;
-import umc.GrowIT.Server.domain.Challenge;
-import umc.GrowIT.Server.domain.enums.ChallengeStatus;
-import umc.GrowIT.Server.service.ChallengeService.ChallengeCommandService;
 import umc.GrowIT.Server.service.ChallengeService.ChallengeQueryService;
 import umc.GrowIT.Server.web.dto.ChallengeDTO.ChallengeRequestDTO;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import umc.GrowIT.Server.apiPayload.ApiResponse;
 import umc.GrowIT.Server.web.dto.ChallengeDTO.ChallengeResponseDTO;
 
-import java.util.List;
-
-@Tag(name = "Challenge", description = "Challenge 관련 API")
+@Tag(name = "Challenge", description = "챌린지 관련 API")
 @RestController
 @RequiredArgsConstructor
 @RequestMapping("/challenges")
-public class ChallengeRestController {
+public class ChallengeController {
     private final ChallengeQueryService challengeQueryService;
     @GetMapping("/summary")
     @Operation(summary = "챌린지 홈 조회 API", description = "사용자의 챌린지 홈 화면에 보여질 챌린지 요약 정보를 조회하는 API입니다.")
