+++ src/main/java/umc/GrowIT/Server/web/controller/KakaoController.java
@@ -1,5 +1,6 @@
 package umc.GrowIT.Server.web.controller;
 
+import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestParam;
@@ -8,6 +9,8 @@ import umc.GrowIT.Server.apiPayload.ApiResponse;
 import umc.GrowIT.Server.web.controller.specification.KakaoSpecification;
 import umc.GrowIT.Server.web.dto.UserDTO.KakaoResponseDTO;
 import umc.GrowIT.Server.service.userService.KakaoService;
+
+@Tag(name = "Kakao", description = "카카오 관련 API")
 @RestController
 @RequiredArgsConstructor
 public class KakaoController implements KakaoSpecification {
