+++ src/main/java/umc/GrowIT/Server/web/controller/TermController.java
@@ -1,5 +1,6 @@
 package umc.GrowIT.Server.web.controller;
 
+import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -10,6 +11,7 @@ import umc.GrowIT.Server.service.termService.TermQueryService;
 
 import java.util.List;
 
+@Tag(name = "Term", description = "약관 동의 관련 API")
 @RestController
 @RequiredArgsConstructor
 public class TermController implements TermSpecification {
